% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notie.R
\name{NotieAlert}
\alias{NotieAlert}
\title{NotieAlert}
\usage{
NotieAlert(message, style_number = 1, duration = 2,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{message}{}

\item{style_number}{}

\item{duration}{}

\item{session}{}
}
\description{
NotieAlert
}

